Port a few rasterizer changes from tuhopuu2: Up the max batch size. Clean up some of the code - move code in headers to source files etc.

+++ source/gameengine/Rasterizer/RAS_MaterialBucket.cpp
@@ -56,7 +56,7 @@ KX_VertexIndex::KX_VertexIndex(int size)
 
 
 
-void KX_VertexIndex::SetIndex(short loc,short index)
+void KX_VertexIndex::SetIndex(short loc,unsigned int index)
 {
 	m_indexarray[loc]=index;
 }
